refactor navigator panel
