+++ src/components/main/actionsPanel/navigatorPanel/components/WorkspacePanel.tsx
@@ -1,0 +1,14 @@
+import React from "react";
+
+export const WorkspacePanel = React.memo(() => {
+  return (
+    <>
+      {/* workspace */}
+      {/* <>
+              <div onClick={onWorkspaceClick}>
+                <img className='icon-s' src={unsavedProject ? (theme === 'Light' ? unsavedLightProjectImg : unsavedDarkProjectImg) : (theme === 'Light' ? projectLightImg : projectDarkImg)}></img>
+              </div>
+            </> */}
+    </>
+  );
+});
