+++ src/components/main/actionsPanel/navigatorPanel/components/index.ts
@@ -1,0 +1,4 @@
+export { WorkspacePanel } from "./WorkspacePanel";
+export { ProjectPanel } from "./ProjectPanel";
+export { DefaultPanel } from "./DefaultPanel";
+export { AdditionalPanel } from "./AdditionalPanel";
