+++ src/components/main/actionsPanel/navigatorPanel/constants.ts
@@ -1,0 +1,6 @@
+export const unsavedDarkProjectImg =
+  "https://rnbw.company/images/favicon-dark-active.png";
+export const unsavedLightProjectImg =
+  "https://rnbw.company/images/favicon-light-active.png";
+export const projectDarkImg = "https://rnbw.company/images/favicon-dark.png";
+export const projectLightImg = "https://rnbw.company/images/favicon-light.png";
