+++ src/components/main/actionsPanel/navigatorPanel/hooks/index.ts
@@ -1,0 +1,2 @@
+export { useNavigationPanelHandlers } from "./useNavigationPanelHandlers";
+export { useFavicon } from "./useFavicon";
